bootloader: Fix progmem attributes
The PROGMEM attribute determines where data is stored, but it has no
influence on how it is accessed, so the welcome message is in fact
copied to RAM at program startup.

Since we lose anyway, we may just stop confusing people with a PGM_P
variable and use plain old (inefficient) const pointers.

BTW this is really only about startup speed. There is always more than
enough RAM for all the static data...