bootloader: inline functions to manipulate the LED
The LED behaviour seems inconsistent. On Arduino boards, the LED is
connected between "pin 13" (PB5) and GND, so it is on when the pin is
high.

OTOH the MONITOR mode assumes inverted logic (a comment says "turn on
LED", followed by clearing the corresponding bit in LED_PORT), so there
might be some boards with the LED connected between "pin 13" and VCC
(but that would leave the LED on when leaving the boot loader?).

Either way, let's leave no doubts about the intention by using an
intuitive function name instead of direct bit manipulation.