bootloader: remove interrupt vector table
IVSEL bit in MCUCR is zero on boot (so interrupt vectors are placed at
the start of the Flash memory), and lock bits specify BLB1 mode 3, so
interrupts are disabled while executing from the Boot Loader section.

Consequently, the interrupt vector table is not used, and its place
can be taken by the initialisation code.

The .text section from crtm168.o must be discarded, because it references
__vector_default, which is in the now-discarded .vectors section.