bootloader: fix asm constraints for scratch counters
Input operands should not be modified by the asm block. If the previous
worked, then it was pure luck. Let's do this properly by declaring the
scratch counters in C.

This commit doesn't change the generated code except allocating other
registers for the counters, because gcc can see that the resulting
counter values are not used for anything.