bootloader: make address an automatic variable
There is no need to keep a global address variable (which must have
static storage), while an automatic variable can be stored in registers.

Note that unlike the static variable, an automatic variable is not
initialised to 0. This should be no problem, because the STK500
Communication Protocol prescribes that the Load Address command
(Cmnd_STK_LOAD_ADDRESS) must always be used prior to Cmnd_STK_PROG_PAGE
or Cmnd_STK_READ_PAGE. If this turns out to be a problem nevertheless,
the address can be initialised to 0 at a cost of 2 instructions.

This commit reduces code size by 60 bytes.