bootloader: build it for ATmega168PA
The Hexbright is an ATmega168PA (which is merely a newer version of
ATmega168P, manufactured using a more advanced technology with smaller
individual transistors).

The ATmega168P (or ATmega168PA) is almost identical to an ATmega168
(or ATmega168A), but it has a different signature byte. This can be
confusing, because the SPI signature is different from the bootloader
signature.