bootloader: use &sym instead of sym for use_pgmvar (non-RAMPZ)
The RAMPZ definition of this macro requires a symbol, not a pointer
(a pointer is only 16 bits, a symbol has all 32 bits in asm), so
the non-RAMPZ definition must do the same.

This hasn't been an issue, because an array identifier in C can be
interpreted as both, but it may cause trouble if you try to use
something else.