bootloader: allow to omit initialization of .data and/or .bss
The current bootloader does not define any initialized data, so the
corresponding loop does nothing. However, simply deleting it from the
file would be too fragile, because if somebody adds any initialized
data, the resulting code would be silently wrong.

Instead, make the initialization of .data optional, and check at link
time that either the .data section is empty, or the appropriate code
was compiled in.

Similarly for the .bss section (but since the .bss section is used by
error_count, the BSS initialization code must be included).