bootloader: update address in prog_buffer
Address auto-increment is broken since moving the code in a separate
function, because changes to address are now local to this function.

Note that since the function gets inlined anyway, the pointer type
does not change the storage, but gcc emits an instruction to move
the target address from register Z to whichever register it uses
for address.