bootloader: fix address increment
The auto-increment is still broken. The address is now correctly
updated when reading or writing, but it is then multiplied by two
on the next read/write.

Since both programmable memory types are word-addressable, let's
convert the address when it is set, so I can get rid of the left
shift later before the actual read or write.