bootloader: implement reading lock bits and fuse bits
The universal SPI programming command can be used to read the value
of lock bits and fuse bits. This is handy if you want to distinguish
between the full Hexbright bootloader and the Hexbright Tiny:

* Hexbright has EFUSE=00.
* Hexbright Tiny has EFUSE=02.

I had to move the definitions for prog_buffer higher up in the file,
so that they could be reused for reading the fuse bits.