bootloader: re-implement app_start() as an inline function
This function pointer is constant (always a jump to location 0), but
it is implemented by gcc as an indirect call, which needs two "ldi"
instructions and one "icall" instruction. But it is possible to
implement this as one "jmp" instruction.

Note that after this change, getch() is considered small enough to be
inlined, but that would increase the bootloader size, so inlining
getch() mustd be disabled explicitly.