More tuning to Mode SOS
Changed the inter-word delay to be an exactly 7 beats and updated the
comments to explain how the bitmap is used to create the dits and dahs.
Also re-ordered all the constants at the top of the program.
Removed unused setting of bit BLOCK_TURNING_OFF as it is never read by
the program.
Move the declaration of time to the start of the program and simple
reset the time at each run through the loop.
Replaced some of the DBG code so it will be compiled in when the
appropriate define is set.
Renamed tailflash to tailflashesLeft because it tells what the variable
is really doing.